+++ NEWS.md
@@ -13,6 +13,7 @@ New features(Analysis):
 + Avoid false positives for the config setting `'assume_real_types_for_internal_functions'`.
   Include all return types for many internal global functions for `--target-php-version` of `7.[0-4]`,
   including those caused by invalid arguments or argument counts.
++ Warn about division, modulo, and exponentiation by 0 (or by values that would cast to 0).
 
 Plugins:
 + If possible, suggest the types that Phan observed during analysis with `UnknownElementTypePlugin`. (#3146)